(bifrost.py, ebysus.py) [REF] no need for two get_varTime methods; EbysusData object will inheret get_varTime from BifrostData.
