+++ helita/sim/ebysus.py
@@ -894,111 +894,7 @@ class EbysusData(BifrostData):
         if var in self.params['aux'][self.snapInd].split():
             return self.get_var(var, *args__get_var, **kw__get_var)
         else:
-            return None
-
-    def get_varTime(self, var, snap=None, iix=None, iiy=None, iiz=None,
-                    mf_ispecies=None, mf_ilevel=None, mf_jspecies=None, mf_jlevel=None,
-                    ifluid=None, jfluid=None, print_freq=None,
-                    *args, **kwargs):
-        """ Gets and returns the value of var over multiple snaps.
-
-        returns the data for the variable (as a 4D array with axes 0,1,2,3 <-> x,y,z,t).
-
-        Parameters
-        ----------
-        var - string
-            Name of the variable to read.
-        snap - list of snapshots, or None (default)
-            Snapshot numbers to read.
-            if None, use self.snap.
-        mf_ispecies - integer, or None (default)
-            Species ID
-            if None, set using other fluid kwargs (see ifluid, iSL, iS).
-            if still None, use self.mf_ispecies
-        mf_ilevel - integer, or None (default)
-            Ionization level
-            if None, set using other fluid kwargs (see ifluid, iSL, iL).
-            if still None, use self.mf_ilevel
-        ifluid - tuple of integers, or None (default)
-            if not None: (mf_ispecies, mf_ilevel) = ifluid
-        print_freq - number, default 2
-            print progress update every print_freq seconds.
-            Use print_freq <= 0 to never print update.
-        **kwargs may contain the following:
-            snaps  - alias for snap
-            iSL    - alias for ifluid
-            jSL    - alias for jfluid
-            iS, iL - alias for ifluid[0], ifluid[1]
-            jS, jL - alias for jfluid[0], jfluid[1]
-        extra **kwargs are passed to get_var, then NOWHERE.
-        extra *args are passed to NOWHERE.
-        """
-
-        # set print_freq
-        if print_freq is None:
-            print_freq = getattr(self, 'print_freq', 2) # default 2
-        else:
-            setattr(self, 'print_freq', print_freq)
-
-        # set snap
-        if snap is None:
-            if 'snaps' in kwargs:
-                snap = kwargs['snaps']
-            if snap is None:
-                snap = self.snap
-        snap = np.array(snap, copy=False)
-        if len(snap.shape)==0:
-            raise ValueError('Expected snap to be list (in get_varTime) but got snap={}'.format(snap))
-        if not np.array_equal(snap, self.snap):
-            self.set_snap(snap)
-            self.variables={}
-
-        # set iix,iiy,iiz; figure out dimensions of return array
-        self.set_domain_iiaxes(iix=iix, iiy=iiy, iiz=iiz)
-
-        snapLen = np.size(self.snap)
-        value = np.empty([self.xLength, self.yLength, self.zLength, snapLen])
-
-        remembersnaps = self.snap                   # remember self.snap (restore later if crash)
-        if hasattr(self, 'recoverData'):
-            delattr(self, 'recoverData')            # smash any existing saved data
-        timestart = now = time.time()               # track timing, so we can make updates.
-        printed_update  = False
-        def _print_clearline(N=100):        # clear N chars, and move cursor to start of line.
-            print('\r'+ ' '*N +'\r',end='') # troubleshooting: ensure end='' in other prints.
-        try:
-            for it in range(0, snapLen):
-                self.snapInd = it
-                # print update if it is time to print update
-                if self.verbose: 
-                    if (print_freq > 0) and (time.time() - now > print_freq):
-                        _print_clearline()
-                        print('Getting {:^10s}; at snap={:2d} (snap_it={:2d} out of {:2d}).'.format(
-                                        var,     snap[it],         it,    snapLen        ), end='')
-                        now = time.time()
-                        print(' Total time elapsed = {:.1f} s'.format(now - timestart), end='')
-                        printed_update=True
-                    
-                # actually get the values here:
-                value[..., it] = self.get_var(var, snap=snap[it],
-                    iix=self.iix, iiy=self.iiy, iiz=self.iiz,
-                    mf_ispecies=mf_ispecies, mf_ilevel=mf_ilevel, ifluid=ifluid,
-                    mf_jspecies=mf_jspecies, mf_jlevel=mf_jlevel, jfluid=jfluid,
-                    **kwargs)
-        except:   # here it is ok to except all errors, because we always raise.
-            if it > 0:
-                self.recoverData = value[..., :it]   # save data 
-                if self.verbose:
-                    print(('Crashed during get_varTime, but managed to get data from {} '
-                           'snaps before crashing. Data was saved and can be recovered '
-                           'via self.recoverData.'.format(it)))
-            raise
-        finally:
-            self.set_snap(remembersnaps)             # restore snaps
-            if printed_update:
-                _print_clearline()
-        
-        return value
+            return None  
 
     def get_nspecies(self):
         return len(self.mf_tabparam['SPECIES'])
